+++ src/Spreads/AssemblyInfo.fs
@@ -4,10 +4,10 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("Spreads")>]
 [<assembly: AssemblyProductAttribute("Spreads")>]
 [<assembly: AssemblyDescriptionAttribute("Spreads")>]
-[<assembly: AssemblyVersionAttribute("0.2.3")>]
-[<assembly: AssemblyFileVersionAttribute("0.2.3")>]
+[<assembly: AssemblyVersionAttribute("0.2.4")>]
+[<assembly: AssemblyFileVersionAttribute("0.2.4")>]
 [<assembly: AssemblyCopyrightAttribute("(c) Victor Baybekov 2015")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "0.2.3"
+    let [<Literal>] Version = "0.2.4"